requirements.txt
django>=2.1,<2.2
django-cors-headers==2.4.0
celery[redis]==4.2.0
djangorestframework-camel-case==1.0.3
drf-dynamic-fields==0.3.0
djangorestframework==3.9.1
uWSGI==2.0.18

# Core dep
openpyxl==2.5.12
pandas==0.23.4
xlrd==1.1.0
ipython==7.2.0
pycairo==1.18.0
PyGObject==3.30.4
